requirements.txt
requests
pandas
azure-ai-textanalytics
python-dotenv
